backend/pyproject.toml
[project]
name = "rag"
version = "0.2.0"
description = "Raccly — NASA bioscience high-precision RAG API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "llama-index-core>=0.14.0",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-llms-openai>=0.6.13",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.35.0",
]
